% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comSplit.R
\name{comSplit}
\alias{comSplit}
\title{Comment split}
\usage{
comSplit(df)
}
\arguments{
\item{df}{The data you wish to divide, containing a Comments column}
}
\description{
Divide a tibble into a list of tibbles by comments. If you dermarcate your runs with comments, this will divide your data into runs.
}
\examples{
licorData(loc,makeCommentsCol=T) \%>\% comSplit()
}
