% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfIncorporate.R
\name{dfIncorporate}
\alias{dfIncorporate}
\title{Incorporate a pulled constant into the dataframe}
\usage{
dfIncorporate(df, newDat, label)
}
\arguments{
\item{df}{The dataframe to incorporate into}

\item{newDat}{The new dataframe to add in, with "value"s and "row"s}

\item{label}{The name of the new column}
}
\description{
Incorporate a pulled constant into the dataframe
}
