% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licorData.R
\name{licorData}
\alias{licorData}
\title{Read in data from the li-6800}
\usage{
licorData(
  location,
  purgeComments = T,
  makeConstCol = F,
  makeCommentsCol = T,
  xlsxIndex = 1
)
}
\arguments{
\item{location}{the path to the data file}

\item{purgeComments}{Removes comments from the file recommended to leave this TRUE - will still work with FALSE but there will be issues}

\item{makeConstCol}{turns S and Oxygen constants into a column. Currently doesn't work with .xlsx files}

\item{makeCommentsCol}{Turns comments into their own column}

\item{xlsxIndex}{which sheet to read on an xlsx.}
}
\description{
This function accepts a licor data file location (xlsx or no-extension (tsv)), and returns a tibble of all the data without the chaff
}
