% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licorFinder.R
\name{licorFinder}
\alias{licorFinder}
\title{Pick the first column that matches what we want}
\usage{
licorFinder(df, colChoose)
}
\arguments{
\item{df}{A dataframe.}

\item{colChoose}{The column we want to find}
}
\description{
Given a dataframe and a column name, find the first column index that contains the chosen name.
}
