% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licorFlashRead.R
\name{licorFlashRead}
\alias{licorFlashRead}
\title{Read licor single flash data jsons
a package, rjson, reads the json data}
\usage{
licorFlashRead(location)
}
\arguments{
\item{location}{The filepath of the json data}
}
\description{
Read licor single flash data jsons
a package, rjson, reads the json data
}
