% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licorFluorReadRCRDG.R
\name{licorFluorReadRCRDG}
\alias{licorFluorReadRCRDG}
\title{A trivial function to read in the RCRDG Li-cor traces}
\usage{
licorFluorReadRCRDG(location)
}
\arguments{
\item{location}{location of the file}
}
\description{
Only works for traces - called "RCRDG"
This data is presented tidy, while single flashes are given true json formatting
}
