% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licorMerge.R
\name{licorMerge}
\alias{licorMerge}
\title{Merge two licor datasets based on time}
\usage{
licorMerge(df1, df2)
}
\arguments{
\item{df1}{The first li-cor data set, from readLicorData::licorData}

\item{df2}{The second li-cor data set, from readLicorData::licorData}
}
\description{
Given two licor datasets, will correct elapsed column based on differences in the hhmmss time.
}
